refactor: rename columns component to config and update import paths
