+++ app/routes/_authenticated+/tasks+/_index/config/columns.tsx
@@ -4,9 +4,9 @@ import { Badge } from '~/components/ui/badge'
 import { Checkbox } from '~/components/ui/checkbox'
 import { labels } from '../../_shared/data/data'
 import type { Task } from '../../_shared/data/schema'
-import { FILTER_FIELD_LABELS } from '../config/constants'
-import { DataTableColumnHeader } from './data-table-column-header'
-import { DataTableRowActions } from './data-table-row-actions'
+import { DataTableColumnHeader } from '../components/data-table-column-header'
+import { DataTableRowActions } from '../components/data-table-row-actions'
+import { FILTER_FIELD_LABELS } from './constants'
 
 export const columns: ColumnDef<Task>[] = [
   {
